Remove assert that non-dynamic GrGpuBuffers are only written once
This was added during a refactor to detect behavior regressions and
in anticipation of removing kStream_GrAccessPattern. That removal
wasn't completed because of skbug.com/11297. Given the advent of
Graphite it likely will never be fixed.

The assert is only detecting whether our behavior conforms to what
is effectively a hint and is making it harder to modify
GrGpuBuffer. It's also a bit of heavy hammer in that for some backends the writes to the buffer are effectively gpu->gpu and so
don't violate the spirit of the hint.

Bug: skia:12720
Change-Id: I27e8d89b63df54a667879f2f7e4452e3c37da264
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/550620
Commit-Queue: Greg Daniel <egdaniel@google.com>
Auto-Submit: Brian Salomon <bsalomon@google.com>
Reviewed-by: Greg Daniel <egdaniel@google.com>
Commit-Queue: Brian Salomon <bsalomon@google.com>