Simplify calculating the reduced font size for transformed glyphs
I'm switching to this two pass method because it will allow
simplifying the buffer structures in MakeInAlloc routine.
I need to simplify the buffer handling to rewrite the interface
to the SkStrikeForGPU to pass only the information the RemoteStrike
has. This will be many CLs to accomplish.

This seems to give the exact same results as the existing code.

Change-Id: I94a31549cbc1924adb1de44e0438996b902fdca8
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/552682
Commit-Queue: Herb Derby <herb@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>