Pass SkGraphcis::GetVariableColrV1Enabled() to FreeType
In [1] SkGraphics was changed to receive a flag callback from Blink if
variable COLRv1 is enabled. Pass this flag down to the corresponding
FreeType property in order to be able to continuously land changes in
FreeType for COLRv1 until shipping it.

Roll FreeType to current head, to include property addition to FreeType
in [2].

[1] https://skia-review.googlesource.com/c/skia/+/551637
[2] https://gitlab.freedesktop.org/freetype/freetype/-/commit/4b6f92e6b3d0d99a839448bc897673b849981997

Bug: skia:13448
Change-Id: I296ec5c5b4093cb1af0300de7d1eadcc8e740a36
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/553436
Reviewed-by: Florin Malita <fmalita@google.com>
Commit-Queue: Dominik Röttsches <drott@google.com>