Fix perspective for drawing of last resort
I don't know why none of the CQ bots don't run scaledemojiperspective,
but this was not found in the gold for the CL. Because it is
clearly wrong.

Take a best guess at the perspective size calculation by using
the matrix's getMaxScale() to estimate the scaling caused by
the perspective matrix.

Change-Id: I58258b18c0c427b39c56f16c7ee17b8b8d4e4e1c
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/553593
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Herb Derby <herb@google.com>