Find the best font size for too big glyphs
If the font produces glyphs that won't fit in the atlas then:

Given the original font as a starting point, find a smaller font
where the maximum glyphs size will fit in the atlas with room for
padding.

From that new font size, calculate a scale factor that will
make the smaller glyphs the same size as the original font.

This is the second attempt at this code.

Change-Id: I2add9a8e9bae59546b7e8c3bbd5c17eb48d90f65
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/554401
Commit-Queue: Herb Derby <herb@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>