Always return a container from SubRunContainer::MakeInAlloc
The old code would return a nullptr on failure or when doing strike
calculations. The rest of the code was not prepared to deal with a
nullptr. When I tried to modify the existing code to deal with a
nullptr, the code became too verbose. I switched to having MakeInAlloc
return an empty SubRunContainer instead of nullptr when performing
strike calculations only.

Bug: oss-fuzz:48695
Bug: oss-fuzz:48690
Change-Id: I633df3316f6f1c3a2b5749722f8c7a0d54ff7f8b
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/556361
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Herb Derby <herb@google.com>