Delay generating SkPaths in PathOpSubmitter
In order to allow the RemoteStrike cache to convert Blobs to Slugs,
the code must create SubRuns using only SkPackedGlyphIDs. Building
PathSubRuns with SkPackedGlyphIDs instead of SkPaths is the first
of 5 conversion.

This code builds a PathSubRun with glyph ids, and delays looking
up the SkPaths until the submitDraw call. The submitDraw uses an
SkStrike which has the facility to generate the path data on demand.

Change-Id: I3270d21b797229117664d10b2f17881afff3c5d9
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/558651
Commit-Queue: Herb Derby <herb@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>