Update child handling in Graphite glue-code generation.
Previously, `emitGlueCodeForEntry` assumed that every child should be
invoked exactly once, and those values should be passed into the shader
snippet.

This model does not work for Runtime Effects. A runtime effect is
free to invoke its children multiple times, or not at all. Now,
glue code generators can choose to call `emit_child_glue_code` and get
the old behavior, or they can handle children their own way as
needed. In practice, GenerateDefaultGlueCode uses this function, but
no other glue-code generators currently use children or have a need
to call it.

This CL also changes `emitGlueCodeForEntry` into a free function
instead of a private method. This makes it possible for us to use
this function from the glue-code generators, which are not part of
SkShaderInfo. SkShaderInfo is documented as a "simple collection
object," so this is probably the right design anyway.

Change-Id: I6975ca92df8a71942f09d7d9ea6f146d9eb3b9ab
Bug: skia:13508
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/558157
Reviewed-by: Robert Phillips <robertphillips@google.com>