Split GenerateDefaultGlueCode into two parts.
We are going to need to treat shaders with children differently than
those without; when children exist, we will need to add a helper
function to the preamble allow all the children to be invoked. This CL
prepares for this change by adding a new
`GenerateDefaultGlueCodeWithChildren` function. At present, it doesn't
change much; it just restructures existing code and updates the
existing snippets to call the new function.

Change-Id: Iae40df25eb1ba82e2e0568bceef5d57994e1deb4
Bug: skia:13508
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/558547
Commit-Queue: Robert Phillips <robertphillips@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>