Emit helper function in GenerateDefaultGlueCode when children exist.
A runtime effect needs to be able to invoke a child snippet from
a `child.eval()` expression anywhere. Because `eval` occurs in an
expression context, we don't have the ability to emit multiple
statements, or declare new variables. That means all of our glue code
needs to eventually become glue-expressions.

When there are no children, the default glue code is just a function
call to a snippet, so this will actually work fine as an expression.
However, when there _are_ children, the glue code becomes arbitrarily
complex. (We need to call child snippets, which could themselves call
more children.)

Now, when children exist, the default glue code emits a helper
function which invokes all of the children first, followed by
invoking the shader snippet passing along all the child outputs.
This gives us some of that Ganesh-like tree structure when
it's needed, but simple flat code when it's not.

Change-Id: I85e42f0794872c9f378ec63581364e0ed55f6047
Bug: skia:13508
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/558308
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>