Replace glue-code generators with expression generators.
A runtime effect needs to be able to invoke a child snippet from
a `child.eval()` expression anywhere. Because `eval` occurs in an
expression context, we don't have the ability to emit multiple
statements. This CL restructures our glue-code generation functions to
return expressions instead of outputting statements, so that they can
be used as subexpressions.

Change-Id: I930bfef2c668323d68f6fd5795621cf42e1bf4eb
Bug: skia:13508
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/558542
Commit-Queue: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>