[bazel] Move Freetype build rules to //bazel/external
One trick about this is that we have the ftoption.h and ftmodule.h
files used to configure how we build Freetype. To make those
available, I had to create a new Bazel workspace exporting those files
(using a symlink to avoid copies). This workspace is located in
//bazel/external/freetype/config and is "imported" by the Skia
workspace as "freetype_config".

This allows Freetype, which is checked out to
//third_party/externals/freetype to use those configuration
include files by means of a genrule that copies them
into the right places.

The BUILD.bazel rules for building Freetype should more closely
match the GN rules [1] (except for brotli support).

[1] https://github.com/google/skia/blob/7d1636010a55664f54165fc6213a37e77609774f/third_party/freetype2/BUILD.gn#L25

Change-Id: Ifc5ebc621c1bed3d65206afc25405cf26585b374
Bug: skia:12541
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/558730
Reviewed-by: Ben Wagner <bungeman@google.com>