[graphite] Report coverage AA from Renderer and respect it in Device
RenderSteps can now report if they emit an analytic coverage variable.
When we have fragment shader integration, this flag can be checked to
handle the fixed blending that happens between the dst color and the
result of the blend between src and dst that is coverage dependent.

Updates Device to take this into account for ordering, anything that
uses coverage AA can depend on the dst so has to be constrained
appropriately. Sets the flag on the text render steps.

This also cleans up some whitespace alignment and stale comments
in Renderer.h

Change-Id: If8cc0bacc6e11166f4ac36e1eec754b371c417b1
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/560727
Reviewed-by: Greg Daniel <egdaniel@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>