[bazel] Move to one big 'kitchen sink' IWYU job
For tractability, we want only one IWYU build that has all the
possible options turned on.

This fixes one place where you could not set multiple GPU backends.
While doing that, I removed the default param from select_multi
because it *needs* to always be set to empty list, otherwise the
"default" case can happen multiple times or unexpectedly.
There is a better way to implement a big "else if none are true"
and that is a different select() statement.

Change-Id: Ib2c95402ec1a75b8ebb541d98b8e5f18bc768415
Bug: skia:13052
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/561897
Owners-Override: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Ben Wagner <bungeman@google.com>