Replace SkSL::Program::Settings with SkSL::ProgramSettings.
Nested types can't be forward declared. Program::Settings was just a
synonym for ProgramSettings that could not be forward declared; it's
now been removed.

Making this change allowed SkSLCompiler.h to forward-declare Program
instead of #including it. This had a ripple effect--many cpps were
implicitly relying on this #include but must explicitly #include
SkSLProgram now.

Change-Id: I8768eae7d09cec854ecf276461bef0c9685336f1
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/562758
Reviewed-by: Arman Uguray <armansito@google.com>