Avoid writing extra copies of builtin functions.
Previously, a call from a runtime effect to a built-in like $hsl_to_rgb
would emit a copy of $hsl_to_rgb_S1, with the pipeline stage suffix
mangled on. This copy wasn't used--writeFunctionCall invokes built-in
functions directly, without the mangle--so it was just generating
cruft. We would later optimize it away as it was legitimately dead
code; this was only visible with optimizations off.

Change-Id: Ic66b79e6cdb3e6f237d6b354e37e748fab94ae18
Bug: skia:13587
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/562762
Commit-Queue: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>