Mark sk_sp with [[clang::trivial_abi]] attribute
Clang's trivial_abi[1] attribute allows types to be marked as safe to
pass in a register instead of on the stack. It also implies that the
type is "trivially relocatable"[2] i.e. a destroying-move operation
(move constructor followed by destructor) is equivalent to a memcpy.

This CL marks sk_sp with the [[clang::trivial_abi]] attribute when the
user configurable SK_SP_TRIVIAL_ABI macro is defined so that it is
considered trivially reloctable. This is so that Chromium PaintOps which
contain sk_sp can be statically asserted to be trivially relocatable so
PaintOpBuffer's memcpy on realloc optimization is guaranteed to be safe,
and potential security issues like crbug.com/1302745 are caught at
compile time.

[1] https://clang.llvm.org/docs/AttributeReference.html#trivial-abi
[2] https://www.open-std.org/jtc1/sc22/wg21/docs/papers/2020/p1144r5.html

Bug: skia:13565, chromium:1302745
Change-Id: Iced60c967da4c8aa6b881e9c23ebb5c914a084c5
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/562257
Reviewed-by: Ben Wagner <bungeman@google.com>
Commit-Queue: Brian Salomon <bsalomon@google.com>
Reviewed-by: Brian Salomon <bsalomon@google.com>
Auto-Submit: Sunny Sachanandani <sunnyps@chromium.org>