Remove expression scanning from function finalization pass.
Now that built-in functions are handled elsewhere, the only thing left
here was a check for calls to `dFdy`. This job is better suited to the
built-in-function scanner anyway.

With this change, the function finalization pass only needs to scan
statements. Expressions can be skipped over entirely.

Change-Id: I265a8708826f5621d905cffcaa00435c0eea6405
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/563124
Commit-Queue: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>