Reduce the snapping tolerance used in SkMatrix::setRotate
SK_ScalarNearlyZero is fairly large, depending on the application. In
this case, it's very close to sin(0.014 degrees), which meant any
rotation smaller than that would be snapped away when setRotate was
called.

We can't remove the snapping entirely, unfortunately. Thanks to
precision (both in converting degrees to radians, and in the trig
functions themselves), we need some region where nearly-zero values are
snapped. This is most important for rotations of 90-degree multiples,
which are now tested. As you increase the number of full rotations, the
"zero" values end up growing (~1e-6 after 5 rotations, ~1e-5 after 40).

The linked bug asks for a rotation of 0.01 degrees, which has a sin of
1.7e-4. The new nearly-zero value (only used for the purposes of trig
function snapping) tries to split the difference, being large enough to
handle up to 40 full rotations before we'll fail to snap correctly,
while allowing *actual* rotations of up to ~0.00087 degrees.

Bug: chromium:1345038
Change-Id: If49d2619e14212b37f3874a0e301d1a04c0e7ed0
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/563460
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>
Reviewed-by: Brian Salomon <bsalomon@google.com>