[bazel] Add skottie_tool_cpu and skottie_tool_gpu rule
Unlike the GN builds of these, the gpu backend is, in fact, not
there in the _cpu version.

Doing a release build of each, we get
skottie_tool_cpu 22M
skottie_tool_gpu 27M

That seems a little big, but it's probably something like
debug symbols still being there. I'll follow up when adding an
"optimize for code size" build flag.

In order to support this, I needed to add the //tools/gpu:utils
rule (an analog for the current GN rule gpu_tool_utils [1]).
I tested this rule on Linux and Mac and translated the other
platforms from the GN rules as best I could, but these are
not yet supported/tested.

I also got rid of the flag "is_skia_dev_build" in favor of
"enable_gpu_test_utils". The former was based on the GN
flag of the same name [2], but is heavily overloaded. I plan
to replace it with more precise, meaningful flags like
"enable_gpu_test_utils".

[1] https://github.com/google/skia/blob/f21a717e1347b022ec72f26dbb8bae66b4c6f7a2/BUILD.gn#L1802
[2] https://github.com/google/skia/blob/f21a717e1347b022ec72f26dbb8bae66b4c6f7a2/BUILD.gn#L94

Bug: skia:12541
Change-Id: Ice652a9cebb68403d09982da1f07409228fa8d7c
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/562936
Reviewed-by: Ben Wagner <bungeman@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>