Remove SkMiniPicture
This was originally slated for deletion in http://review.skia.org/214661
but was reverted due to use in Flutter. AFAICT, Flutter no longer uses
this.

I moved bool SkPicture_StreamIsSKP(SkStream*, SkPictInfo*) to
SkPicturePriv.h, as that was the only thing left in SkPictureCommon.h
(SkPicturePriv.h did not exist at the time of the original removal CL).

From the original removal CL:
> This was an optimization for Chromium that I believe is no longer
> relevant in a world of PaintOpBuffers.

Change-Id: Ie6972a6fd76f5bd2b7b6d429e26cb9e83b028d93
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/564896
Reviewed-by: Brian Salomon <bsalomon@google.com>