Change to using SkPackedGlyphID and left top point
Currently, glyph data is communicated to the SubRuns using
SkGlyph* and the glyph origin. In the near future SkGlyph* will
not be an option to pass data to the SubRuns. Switch from using
SkGlyph* to using SkPackedGlyphID and from using the glyph
origin to using the top left corner of the glyph bounding
rectangle.

Change-Id: I912e07312429962067241c13df5b81072f49eefd
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/565642
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Herb Derby <herb@google.com>