[bazel] Add CanvasKit build with CPU-backend only
This adds a small change to the test harness that allows
skipped tests (e.g. gms that exercises GPU specific code) to
not upload anything to Gold. Previously, they would upload
a blank canvas, which was marked "positive" because the CPU
backend would not support this.

The CPU build still needs the legacy shader context for
perlin noise drawing.

Change-Id: I3ee9cbd8b9538367bad673250965b9d7c7fa311a
Bug: skia:12541
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/565014
Reviewed-by: Brian Osman <brianosman@google.com>