Propagate SkSurfaceProps in SkPictureShader
Currently when an `SkPictureShader` is rendered, it always uses the
default `SkSurfaceProps`, preventing any non-default flags from being
used when rendering text inside the shader. (This is a very minor issue,
currently, but new SkSurfaceProps flags/properties are planned for the
future.)

This change propagates the `SkSurfaceProps` from the target surface,
through several layers of function calls, to
`SkPictureShader::rasterShader` and `asFragmentProcessor`, where it is
needed.  To limit the number of changed functions signatures,
SkSurfaceProps were added to `GrFPArgs`, `SkDraw`, `SkStageRec`, and
`SkShaderBase::ContextRec`.

Bug: skia:13369
Change-Id: I086bf7ca9732a6e5ab3a617d60c9c1708a4c5b18
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/557736
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>