Have DrawableSubRuns take SkPackedGlyphIDs instead of SkGlyph*
This is the last subrun type to convert over. It is very similar
to the conversion done for PathSubRuns.

An odd feature of the existing code is that the ownership of the
SkDrawables is done in bulk through the SkStrike. The ownership
is a little more explicit now.

Change-Id: Iea7e6f51cdb24e502c85734c40ac836ea23a83ef
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/566556
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Herb Derby <herb@google.com>