[codesize] Strip out debug symbols before bloaty runs
This adds a new CIPD package for a subset of binutils
(we want "strip" which removes debug symbols). I'm not sure
if this is the best approach because binutils appears to
have some extra dynamic library deps that might not work
when we update our swarming machines. We might need to bundle
those deps into the CIPD package.

Before: http://screen/98JoTV9endfUZQi
After: http://screen/6Ym4WszZ7kZbSbX

Change-Id: Ic97511c44c45c641b270e48593e9d0bebf659e3e
Bug: skia:13620
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/565901
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>